+++ packages/@jsii/go-runtime/jsii-experimental/runtime.go
@@ -4,8 +4,14 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
+	"reflect"
 	"regexp"
 )
+
+// Maps interface types to their concrete implementation structs. Used by
+// `castAndSetToPtr` to instantiate a concrete type that implements the
+// the interface as dictated by the type of the ret value.
+type implementationMap = map[reflect.Type]reflect.Type
 
 // Load ensures a npm package is loaded in the jsii kernel.
 func Load(name string, version string, tarball []byte) {
@@ -38,12 +44,13 @@ func Load(name string, version string, tarball []byte) {
 
 // Create will construct a new JSII object within the kernel runtime. This is
 // called by jsii object constructors.
-func Create(fqn FQN, args []interface{}, interfaces []FQN, overrides []Override, returns interface{}) {
+func Create(fqn FQN, args []interface{}, interfaces []FQN, overrides []Override, ret interface{}) {
 	client := getClient()
+
 	res, err := client.create(createRequest{
 		Api:        "create",
 		Fqn:        fqn,
-		Args:       args,
+		Args:       castPtrsToRef(args),
 		Interfaces: interfaces,
 		Overrides:  overrides,
 	})
@@ -52,12 +59,12 @@ func Create(fqn FQN, args []interface{}, interfaces []FQN, overrides []Override,
 		panic(err)
 	}
 
-	client.objects[returns] = res.JsiiInstanceId
+	client.objects[ret] = res.JsiiInstanceId
 }
 
 // Invoke will call a method on a jsii class instance. The response should be
 // decoded into the expected return type for the method being called.
-func Invoke(obj interface{}, method string, args []interface{}, returns interface{}) {
+func Invoke(obj interface{}, method string, args []interface{}, hasReturn bool, ret interface{}, implMap implementationMap) {
 	client := getClient()
 
 	// Find reference to class instance in client
@@ -67,10 +74,10 @@ func Invoke(obj interface{}, method string, args []interface{}, returns interfac
 		panic("No Object Found")
 	}
 
-	_, err := client.invoke(invokeRequest{
+	res, err := client.invoke(invokeRequest{
 		Api:    "invoke",
 		Method: method,
-		Args:   args,
+		Args:   castPtrsToRef(args),
 		Objref: objref{
 			JsiiInstanceId: refid,
 		},
@@ -79,24 +86,32 @@ func Invoke(obj interface{}, method string, args []interface{}, returns interfac
 	if err != nil {
 		panic(err)
 	}
+
+	if hasReturn {
+		castAndSetToPtr(ret, res.Result, implMap)
+	}
 }
 
-func InvokeStatic(fqn FQN, method string, args []interface{}, returns interface{}) {
+func InvokeStatic(fqn FQN, method string, args []interface{}, hasReturn bool, ret interface{}, implMap implementationMap) {
 	client := getClient()
 
-	_, err := client.sinvoke(staticInvokeRequest{
+	res, err := client.sinvoke(staticInvokeRequest{
 		Api:    "sinvoke",
 		Fqn:    fqn,
 		Method: method,
-		Args:   args,
+		Args:   castPtrsToRef(args),
 	})
 
 	if err != nil {
 		panic(err)
 	}
+
+	if hasReturn {
+		castAndSetToPtr(ret, res.Result, implMap)
+	}
 }
 
-func Get(property string, obj interface{}) {
+func Get(obj interface{}, property string, ret interface{}, implMap implementationMap) {
 	client := getClient()
 
 	// Find reference to class instance in client
@@ -106,23 +121,25 @@ func Get(property string, obj interface{}) {
 		panic("No Object Found")
 	}
 
-	_, err := client.get(getRequest{
+	res, err := client.get(getRequest{
 		Api:      "get",
 		Property: property,
-		Objref:   objref{JsiiInstanceId: refid},
+		Objref: objref{
+			JsiiInstanceId: refid,
+		},
 	})
 
 	if err != nil {
 		panic(err)
 	}
 
-	// Do we need to return Value from getResponse?
+	castAndSetToPtr(ret, res.Value, implMap)
 }
 
-func StaticGet(fqn FQN, property string) {
+func StaticGet(fqn FQN, property string, ret interface{}, implMap implementationMap) {
 	client := getClient()
 
-	_, err := client.sget(staticGetRequest{
+	res, err := client.sget(staticGetRequest{
 		Api:      "sget",
 		Fqn:      fqn,
 		Property: property,
@@ -131,9 +148,11 @@ func StaticGet(fqn FQN, property string) {
 	if err != nil {
 		panic(err)
 	}
+
+	castAndSetToPtr(ret, res.Value, implMap)
 }
 
-func Set(property string, value, obj interface{}) {
+func Set(obj interface{}, property string, value interface{}) {
 	client := getClient()
 
 	// Find reference to class instance in client
@@ -146,7 +165,10 @@ func Set(property string, value, obj interface{}) {
 	_, err := client.set(setRequest{
 		Api:      "set",
 		Property: property,
-		Objref:   objref{JsiiInstanceId: refid},
+		Value:    castPtrToRef(value),
+		Objref: objref{
+			JsiiInstanceId: refid,
+		},
 	})
 
 	if err != nil {
@@ -166,6 +188,96 @@ func StaticSet(fqn FQN, property string, value interface{}) {
 
 	if err != nil {
 		panic(err)
+	}
+}
+
+func castValToRef(data interface{}) (objref, bool) {
+	ref := objref{}
+	ok := false
+	dataVal := reflect.ValueOf(data)
+
+	if dataVal.Kind() == reflect.Map {
+		for _, k := range dataVal.MapKeys() {
+			// Finding values type requires extracting from reflect.Value
+			// otherwise .Kind() returns `interface{}`
+			v := reflect.ValueOf(dataVal.MapIndex(k).Interface())
+
+			if k.Kind() == reflect.String && k.String() == "$jsii.byref" && v.Kind() == reflect.String {
+				ref.JsiiInstanceId = v.String()
+				ok = true
+			}
+
+		}
+	}
+
+	return ref, ok
+}
+
+// Accepts pointers to structs that implement interfaces and searches for an
+// existing object reference in the client. If it exists, it casts it to an
+// objref for the runtime. Recursively casts types that may contain nested
+// object references.
+func castPtrToRef(data interface{}) interface{} {
+	client := getClient()
+	dataVal := reflect.ValueOf(data)
+
+	if dataVal.Kind() == reflect.Ptr {
+		valref, valHasRef := client.findObjectRef(data)
+		if valHasRef {
+			return objref{JsiiInstanceId: valref}
+		}
+	} else if dataVal.Kind() == reflect.Slice {
+		refs := make([]interface{}, dataVal.Len())
+		for i := 0; i < dataVal.Len(); i++ {
+			refs[i] = dataVal.Index(i).Interface()
+		}
+		return refs
+	}
+
+	return data
+}
+
+// Casts slice of data into new slice of data with pointers to interfaces
+// converted to objrefs. This is useful for casting arguments to methods and
+// constructors to data that can be serialized before being passed over the
+// wire.
+func castPtrsToRef(args []interface{}) []interface{} {
+	argRefs := make([]interface{}, len(args))
+	for i, arg := range args {
+		argRefs[i] = castPtrToRef(arg)
+	}
+
+	return argRefs
+}
+
+// castAndSetToPtr accepts a pointer to any type and attempts to cast the value
+// argument to be the same type. Then it sets the value of the pointer element
+// to be the newly cast data. This is used to cast payloads from JSII to
+// expected return types for Get and Invoke functions.
+func castAndSetToPtr(ptr interface{}, data interface{}, implMap implementationMap) {
+	ptrVal := reflect.ValueOf(ptr).Elem()
+	dataVal := reflect.ValueOf(data)
+
+	ref, isRef := castValToRef(data)
+
+	if ptrVal.Kind() == reflect.Slice && dataVal.Kind() == reflect.Slice {
+		// If return type is a slice, recursively cast elements
+		for i := 0; i < dataVal.Len(); i++ {
+			innerType := ptrVal.Type().Elem()
+			inner := reflect.New(innerType)
+
+			castAndSetToPtr(inner.Interface(), dataVal.Index(i).Interface(), implMap)
+			ptrVal.Set(reflect.Append(ptrVal, inner.Elem()))
+		}
+	} else if isRef {
+		// If return data is JSII object references, add to objects table.
+		concreteType := implMap[ptrVal.Type()]
+		ptrVal.Set(reflect.New(concreteType))
+		client := getClient()
+		client.objects[ptrVal.Interface()] = ref.JsiiInstanceId
+	} else {
+		val := reflect.ValueOf(data)
+		ptrVal.Set(val)
 	}
 }