+++ packages/jsii-pacmak/lib/targets/go/package.ts
@@ -14,6 +14,9 @@ import {
 import { GoClass, GoType, Enum, Interface, Struct } from './types';
 import { findTypeInTree, goPackageName, flatMap } from './util';
 
+export const GOMOD_FILENAME = 'go.mod';
+export const GO_VERSION = '1.15';
+
 /*
  * Package represents a single `.go` source file within a package. This can be the root package file or a submodule
  */
@@ -66,18 +69,22 @@ export abstract class Package {
   }
 
   /*
-   * The module names of this modules dependencies. Used for import statements
+   * goModuleName returns the full path to the module name.
+   * Used for import statements and go.mod generation
+   */
+  public get goModuleName(): string {
+    const moduleName = this.root.moduleName;
+    const prefix = moduleName !== '' ? `${moduleName}/` : '';
+    const rootPackageName = this.root.packageName;
+    const suffix = this.filePath !== '' ? `/${this.filePath}` : ``;
+    return `${prefix}${rootPackageName}${suffix}`;
+  }
+
+  /*
+   * The module names of this module's dependencies. Used for import statements.
    */
   public get dependencyImports(): Set<string> {
-    return new Set(
-      this.dependencies.map((pack) => {
-        const moduleName = pack.root.moduleName;
-        const prefix = moduleName !== '' ? `${moduleName}/` : '';
-        const rootPackageName = pack.root.packageName;
-        const suffix = pack.filePath !== '' ? `/${pack.filePath}` : ``;
-        return `${prefix}${rootPackageName}${suffix}`;
-      }),
-    );
+    return new Set(this.dependencies.map((pkg) => pkg.goModuleName));
   }
 
   /*
@@ -89,6 +96,7 @@ export abstract class Package {
 
   public emit(context: EmitContext): void {
     const { code } = context;
+
     code.openFile(this.file);
     this.emitHeader(code);
     this.emitImports(code);
@@ -123,10 +131,27 @@ export abstract class Package {
     );
   }
 
+  protected get usesReflectionPackage(): boolean {
+    return (
+      this.types.some((type) => type.usesReflectionPackage) ||
+      this.submodules.some((sub) => sub.usesReflectionPackage)
+    );
+  }
+
   private emitImports(code: CodeMaker) {
     code.open('import (');
     if (this.usesRuntimePackage) {
       code.line(`${JSII_RT_ALIAS} "${JSII_RT_MODULE_NAME}"`);
+    }
+
+    if (this.usesInitPackage) {
+      code.line(
+        `${JSII_INIT_ALIAS} "${this.root.moduleName}/${this.root.packageName}/${JSII_INIT_PACKAGE}"`,
+      );
+    }
+
+    if (this.usesReflectionPackage) {
+      code.line(`"reflect"`);
     }
 
     for (const packageName of this.dependencyImports) {
@@ -134,12 +159,6 @@ export abstract class Package {
       if (packageName !== this.packageName) {
         code.line(`"${packageName}"`);
       }
-    }
-
-    if (this.usesInitPackage) {
-      code.line(
-        `${JSII_INIT_ALIAS} "${this.root.moduleName}/${this.root.packageName}/${JSII_INIT_PACKAGE}"`,
-      );
     }
 
     code.close(')');
@@ -160,6 +179,7 @@ export abstract class Package {
  */
 export class RootPackage extends Package {
   public readonly assembly: Assembly;
+  public readonly version: string;
   private readonly readme?: ReadmeFile;
 
   public constructor(assembly: Assembly) {
@@ -176,6 +196,7 @@ export class RootPackage extends Package {
     );
 
     this.assembly = assembly;
+    this.version = assembly.version;
 
     if (this.assembly.readme?.markdown) {
       this.readme = new ReadmeFile(
@@ -189,6 +210,23 @@ export class RootPackage extends Package {
     super.emit(context);
     this.emitJsiiPackage(context);
     this.readme?.emit(context);
+
+    this.emitGomod(context.code);
+  }
+
+  private emitGomod(code: CodeMaker) {
+    code.openFile(GOMOD_FILENAME);
+    code.line(`module ${this.goModuleName}`);
+    code.line();
+    code.line(`go ${GO_VERSION}`);
+    code.line();
+    code.open('require (');
+    code.line(`${JSII_RT_MODULE_NAME} v${this.assembly.jsiiVersion}`);
+    for (const dep of this.packageDependencies) {
+      code.line(`${dep.goModuleName} v${dep.version}`);
+    }
+    code.close(')');
+    code.closeFile(GOMOD_FILENAME);
   }
 
   /*
@@ -245,10 +283,12 @@ export class RootPackage extends Package {
         );
       }
     }
+
     code.close(')');
     code.line();
     code.line('var once sync.Once');
     code.line();
+
     code.line(
       `// ${JSII_INIT_FUNC} performs the necessary work for the enclosing`,
     );
@@ -268,6 +308,7 @@ export class RootPackage extends Package {
     );
     code.close('})');
     code.close('}');
+
     code.closeFile(file);
   }
 }