+++ packages/jsii-pacmak/lib/targets/go/types/interface.ts
@@ -1,40 +1,54 @@
-import { toPascalCase } from 'codemaker';
 import { InterfaceType, Method, Property } from 'jsii-reflect';
 
 import { EmitContext } from '../emit-context';
 import { Package } from '../package';
+import { MethodCall } from '../runtime';
 import { getMemberDependencies, getParamDependencies } from '../util';
-import { GoType } from './go-type';
+import { GoType, INTERFACE_TYPE_SUFFIX } from './go-type';
 import { GoTypeRef } from './go-type-reference';
-import { GoMethod, GoTypeMember } from './type-member';
+import { GoMethod, GoProperty } from './type-member';
 
 export class Interface extends GoType {
   public readonly methods: InterfaceMethod[];
+  public readonly inheritedMethods: InterfaceMethod[];
   public readonly properties: InterfaceProperty[];
+  public readonly inheritedProperties: InterfaceProperty[];
+  public readonly interfaceName: string;
 
   public constructor(pkg: Package, public type: InterfaceType) {
     super(pkg, type);
     this.methods = Object.values(type.getMethods()).map(
       (method) => new InterfaceMethod(this, method),
     );
+
+    this.inheritedMethods = Object.values(type.getMethods(true)).map(
+      (method) => new InterfaceMethod(this, method),
+    );
+
     this.properties = Object.values(type.getProperties()).map(
       (prop) => new InterfaceProperty(this, prop),
     );
+
+    this.inheritedProperties = Object.values(type.getProperties(true)).map(
+      (prop) => new InterfaceProperty(this, prop),
+    );
+
+    this.interfaceName = `${this.name}${INTERFACE_TYPE_SUFFIX}`;
   }
 
   public emit(context: EmitContext) {
     this.emitDocs(context);
 
     const { code } = context;
-    code.openBlock(`type ${code.toPascalCase(this.name)} interface`);
+    code.openBlock(`type ${this.interfaceName} interface`);
 
     // embed extended interfaces
     for (const iface of this.extends) {
-      code.line(iface.scopedName(this.pkg));
+      code.line(iface.scopedInterfaceName(this.pkg));
     }
 
     for (const method of this.methods) {
-      method.emit(context);
+      method.emitDecl(context);
     }
 
     for (const prop of this.properties) {
@@ -43,6 +57,21 @@ export class Interface extends GoType {
 
     code.closeBlock();
     code.line();
+
+    code.line(`type ${this.name} struct {}`);
+    code.line();
+
+    for (const method of this.inheritedMethods) {
+      method.emit(context);
+    }
+
+    for (const prop of this.inheritedProperties) {
+      prop.emitGetterImpl(context);
+
+      if (!prop.immutable) {
+        prop.emitSetterImpl(context);
+      }
+    }
   }
 
   public get usesInitPackage() {
@@ -57,6 +86,10 @@ export class Interface extends GoType {
       this.properties.some((p) => p.usesRuntimePackage) ||
       this.methods.some((m) => m.usesRuntimePackage)
     );
+  }
+
+  public get usesReflectionPackage(): boolean {
+    return this.properties.length > 0 || this.methods.length > 0;
   }
 
   public get extends(): GoTypeRef[] {
@@ -87,29 +120,19 @@ export class Interface extends GoType {
   }
 }
 
-class InterfaceProperty implements GoTypeMember {
-  public readonly name: string;
-  public readonly getter: string;
+class InterfaceProperty extends GoProperty {
   public readonly reference?: GoTypeRef;
-
-  public readonly usesInitPackage = false;
-  public readonly usesRuntimePackage = false;
 
   public constructor(
     public readonly parent: Interface,
-    private readonly property: Property,
+    public readonly property: Property,
   ) {
-    this.name = toPascalCase(property.name);
-    this.getter = `Get${this.name}`;
-
-    if (property.type) {
-      this.reference = new GoTypeRef(parent.pkg.root, property.type);
-    }
+    super(parent, property);
   }
 
   public get returnType(): string {
     return (
-      this.reference?.scopedName(this.parent.pkg) ??
+      this.reference?.scopedInterfaceName(this.parent.pkg) ??
       this.property.type.toString()
     );
   }
@@ -126,22 +149,42 @@ class InterfaceProperty implements GoTypeMember {
 }
 
 class InterfaceMethod extends GoMethod {
+  public readonly runtimeCall: MethodCall;
   public readonly usesInitPackage = false;
-  public readonly usesRuntimePackage = false;
+  public readonly usesRuntimePackage = true;
 
   public constructor(
     public readonly parent: Interface,
     public readonly method: Method,
   ) {
     super(parent, method);
+    this.runtimeCall = new MethodCall(this);
   }
 
-  public emit(context: EmitContext) {
+  public emitDecl(context: EmitContext) {
     const docs = this.method.docs;
     if (docs) {
       context.documenter.emit(docs);
     }
     const { code } = context;
-    code.line(`${this.name}(${this.paramString()})${this.returnType}`);
+    code.line(`${this.name}(${this.paramString()})${this.returnTypeString}`);
+  }
+
+  public emit({ code }: EmitContext) {
+    const name = this.name;
+    code.openBlock(
+      `func (${this.instanceArg} *${
+        this.parent.name
+      }) ${name}(${this.paramString()})${this.returnTypeString}`,
+    );
+
+    this.runtimeCall.emit(code);
+
+    code.closeBlock();
+    code.line();
+  }
+
+  private get returnTypeString(): string {
+    return this.reference?.void ? '' : ` ${this.returnType}`;
   }
 }