+++ packages/jsii-pacmak/lib/targets/python.ts
@@ -6,6 +6,7 @@ import * as reflect from 'jsii-reflect';
 import {
   Translation,
   Rosetta,
+  enforcesStrictMode,
   typeScriptSnippetFromSource,
 } from 'jsii-rosetta';
 import * as path from 'path';
@@ -2169,7 +2170,11 @@ class PythonGenerator extends Generator {
   }
 
   public convertExample(example: string): string {
-    const snippet = typeScriptSnippetFromSource(example, 'example');
+    const snippet = typeScriptSnippetFromSource(
+      example,
+      'example',
+      enforcesStrictMode(this.assembly),
+    );
     const translated = this.rosetta.translateSnippet(snippet, 'python');
     if (!translated) {
       return example;
@@ -2181,6 +2186,7 @@ class PythonGenerator extends Generator {
     return this.rosetta.translateSnippetsInMarkdown(
       markdown,
       'python',
+      enforcesStrictMode(this.assembly),
       (trans) => ({
         language: trans.language,
         source: this.prefixDisclaimer(trans),