+++ packages/jsii-rosetta/test/translations/classes/constructor_with_optional_params.java
@@ -7,6 +7,6 @@ public class A {
         this(a, 3);
     }
 
-    public A(String a, int b) {
+    public A(String a, Number b) {
     }
 }